added employee classes and made changes to currency values
Added all the employee classes with skeleton code and no comments. Made changes to the currency values from float to double to match real world use

+++ Perfect Event Agency/src/Booking.java
@@ -3,7 +3,7 @@ public class Booking {
     private Venue venue;
     private FoodSelection foodSelection;
     private Options options;
-    private int totalCost;
+    private double totalCost;
 
     public Booking() {
     };
@@ -14,6 +14,7 @@ public class Booking {
         this.foodSelection = foodSelection;
         this.options = options;
         this.totalCost = 0;
+
     }
 
     //getter and setter methods
@@ -52,7 +53,7 @@ public class Booking {
         this.options = options;
     }
 
-    public int getTotalCost() {
+    public double getTotalCost() {
         return totalCost;
     }
